+++ src/geometry/sphere.h
@@ -4,19 +4,18 @@
 #include "vector.h"
 #include "ray.h"
 #include "material.h"
-#include "math_helper.h"
 
 
 class Sphere
 {
     public:
         std::string name;
-        float radius;
+        double radius;
         Vector3 position;
         BSDF material;
 
         Sphere();
-        Sphere(const std::string& tempName, float tempRadius, const Vector3& tempPosition, const BSDF& tempMaterial);
+        Sphere(const std::string& tempName, double tempRadius, const Vector3& tempPosition, const BSDF& tempMaterial);
 
         float computeIntersection(const Ray& ray);
